Icons etc. identified, start of reversing ROM-embedded resources.
I recycled a Lua script from reversing an 8-bit game to convert the
ROM into a bitmap, and searched it for images.

The ROM contains some extremely large dithered images of the dev team,
but it looks like they're RLE or something, so they don't display
directly. I will need to reverse them.

I also discovered that a list of resource types was part of a resource
file, as documented by Inside Macintosh, so I need to do a bit more
reversing there to fully extract the resources.

I also discovered a bit more 68K assembly that hadn't been
identified. It feels like we're getting there!